Update yousave.py, Create subsave.py
Added the ability to save data inside subtitles on youtube. Originally this was done in hex like yousave, however I figured it would be better to transfer this into base64 (segments of 20 lines). Overall, I am pretty impressed with how it works, and am looking forward to using it more.